+++ gateway/src/main/java/org/acumos/federation/gateway/config/LoggingConfiguration.java
@@ -1,0 +1,50 @@
+/*-
+ * ===============LICENSE_START=======================================================
+ * Acumos
+ * ===================================================================================
+ * Copyright (C) 2019 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
+ * ===================================================================================
+ * This Acumos software file is distributed by AT&T and Tech Mahindra
+ * under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *  
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ * This file is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ===============LICENSE_END=========================================================
+ */
+
+package org.acumos.federation.gateway.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+/**
+ * Adds logging handler interceptor. <BR>
+ * http://www.devgrok.com/2017/04/adding-mdc-headers-to-every-spring-mvc.html
+ */
+@Configuration
+public class LoggingConfiguration {
+
+	@Bean
+	public LoggingHandlerInterceptor loggingHandlerInterceptor() {
+		return new LoggingHandlerInterceptor();
+	}
+
+	@Bean
+	public WebMvcConfigurer webConfigurer() {
+		return new WebMvcConfigurer() {
+			@Override
+			public void addInterceptors(InterceptorRegistry registry) {
+				registry.addInterceptor(loggingHandlerInterceptor());
+			}
+		};
+	}
+
+}